Simplify clipping and cleanup history files.
When doing a scatter plot, only plot points that fall into the specified
window range. Previously, points to the left of the left y-axis would
be erroneously plotted. This is now fixed.

Also, ensure that restarts (and backups, I guess) are removed from
history data. This is done by going through the model number column
in reverse and ensuring that each previous model number is indeed less
than the one after it. If not, the row is removed.