Fix x-axis labels in dark mode and downloading.
With new color mode switching, the x-axis labels were not being
updated to the correct color. This commit fixes that issue. Now
whenever the color mode is switched, the x-axis color is changed
dynamically, but when saving a plot in dark mode, the x-axis label
is forced to be black.